Create cog.py
This code defines a list of words, shuffles them randomly, and then displays each word one at a time, asking the user to type in the word they see. After the user enters their answer, the code checks if it matches the displayed word and prints out a message indicating whether the user was correct or not. The code also pauses for 1 second between each word to give the user time to respond. This simple program could be used to assess working memory and attention abilities.